kafka: move none public/common parts to kafka-internal.c
