+++ modules/kafka/Makefile.am
@@ -17,6 +17,7 @@ modules_kafka_libkafka_la_SOURCES = \
   modules/kafka/kafka-dest-worker.h \
   modules/kafka/kafka-dest-worker.c \
   modules/kafka/kafka-internal.h \
+  modules/kafka/kafka-internal.c \
   modules/kafka/kafka-plugin.c
 
 modules_kafka_libkafka_la_CPPFLAGS  = \
