+++ modules/kafka/kafka-dest-worker.c
@@ -23,7 +23,6 @@
  *
  */
 #include "kafka-dest-worker.h"
-#include "kafka-dest-driver.h"
 #include "kafka-internal.h"
 #include "str-utils.h"
 #include "timeutils/misc.h"
